// tina-graphql-gateway/examples/tina-tailwind-sidebar-demo/.tina/__generated__/types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: any }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** References another document, used as a foreign key */
  Reference: any;
  JSON: any;
  JSONObject: any;
};
// ...
export type Nav_Items_Data = {
  __typename?: 'Nav_Items_Data';
  label?: Maybe<Scalars['String']>;
  link?: Maybe<Scalars['String']>;
};

export type Nav_Data = {
  __typename?: 'Nav_Data';
  title?: Maybe<Scalars['String']>;
  items?: Maybe<Array<Maybe<Nav_Items_Data>>>;
};

export type Homepage_Navlist_Data = Nav_Data;

export type HomepageFooter_Social_Data = {
  __typename?: 'HomepageFooter_Social_Data';
  facebook?: Maybe<Scalars['String']>;
  twitter?: Maybe<Scalars['String']>;
  instagram?: Maybe<Scalars['String']>;
  github?: Maybe<Scalars['String']>;
};

export type Homepage_Footer_Data = {
  __typename?: 'Homepage_Footer_Data';
  social?: Maybe<HomepageFooter_Social_Data>;
};